requirements.txt
streamlit==1.56.0
pandas
psycopg2-binary
plotly==6.7.0